third_party/WebKit/LayoutTests/fast/css/absolute-position-with-negative-width.html

<!DOCTYPE HTML>
<script src="../../resources/js-test.js"></script>

<div style="position: relative; width: 0px">
    <div style="position: absolute; left: 1px; right: 1px;" id="absolute_with_left_right"></div>
</div>

<script>
absolute_with_left_right = document.getElementById("absolute_with_left_right");
shouldBe("absolute_with_left_right.getBoundingClientRect().width","0");
shouldBe("absolute_with_left_right.getClientRects()[0].width","0");
shouldBe("absolute_with_left_right.offsetWidth","0");
</script>
